example

example

Get Started

Internal functions

Some Theory in JWAS

A Table for Bayesian Linear Mixed Models (BLMM)

Models

Bayesian Linear Additive Genetic Model

Univariate Linear Additive Genetic Model

Multivariate Linear Additive Genetic Model

guide

Bayesian Linear Mixed Models (Genomic Data)

Univariate Linear Mixed Model (Genomic data)

Multivariate Linear Mixed Model (Genomic data)

Bayesian Linear Mixed Models

Univariate Linear Mixed Model (conventional)

Multivariate Linear Mixed Model (conventional)

Supporting and Citing

Get Started:

Public functions

example

Get Started

example 1

example 2

example 1

JWAS.runMCMC — Function.

runMCMC(mme,df;Pi=0.0,estimatePi=false,chain_length=1000,starting_value=false,printout_frequency=100,missing_phenotypes=false,constraint=false,methods="conventional (no markers)",output_samples_frequency::Int64 = 0)

Run MCMC (marker information included or not) with sampling of variance components.

available methods include "conventional (no markers)", "BayesC0", "BayesC", "BayesCC","BayesB".

missing_phenotypes

Pi for single-trait analyses is a number; Pi for multi-trait analyses is a dictionary such as Pi=Dict([1.0; 1.0]=>0.7,[1.0; 0.0]=>0.1,[0.0; 1.0]=>0.1,[0.0; 0.0]=>0.1),

if Pi (Π) is not provided in multi-trait analysis, it will be generated assuming all markers have effects on all traits.

save MCMC samples every output_samples_frequency iterations

starting_value can be provided as a vector for all location parameteres except marker effects.

print out the monte carlo mean in REPL with printout_frequency

constraint=true if constrain residual covariances between traits to be zero.

source

link to JWAS.jl Documentation

link to add_genotypes

example 2

JWAS.add_genotypes — Function.

add_genotypes(mme::MME,file,G;separator=' ',header=true,center=true,G_is_marker_variance=false,df=4.0)

Get marker informtion from a genotype file (same order as the phenotype file).

G defaults to the genetic variance with degree of freedom df=4.0.

File format:

Animal,marker1,marker2,marker3,marker4,marker5
S1,1,0,1,1,1
D1,2,0,2,2,1
O1,1,2,0,1,0
O3,0,0,2,1,1

source

JWAS.add_markers — Function.

same to add_genotypes

source

Test these examples