Internals

Internal functions

Documentation for JWAS.jl's internal (private) interface, which are not available to general users. These internal functions are small blocks that public function build on.

<!–-

Index

JWAS.PedModule.get_pedigree

JWAS.add_genotypes

JWAS.add_markers

JWAS.add_term

JWAS.build_model

JWAS.getEBV

JWAS.getEBV

JWAS.getEBV

JWAS.getMME

JWAS.get_outputX_others

JWAS.outputEBV

JWAS.outputMCMCsamples

JWAS.runMCMC

JWAS.set_covariate

JWAS.set_random

JWAS.set_random

JWAS.showMME

JWAS.solve

Internal interface

JWAS.add_genotypes — Method.

add_genotypes(mme::MME,file,G;separator=' ',header=true,center=true,G_is_marker_variance=false,df=4.0)

Get marker informtion from a genotype file.

G defaults to the genetic variance with degree of freedom df=4.0.

File format:

Animal,marker1,marker2,marker3,marker4,marker5
S1,1,0,1,1,1
D1,2,0,2,2,1
O1,1,2,0,1,0
O3,0,0,2,1,1

source

JWAS.add_markers — Method.

same to add_genotypes

source

JWAS.build_model — Method.

build_model(model_equations::AbstractString,R;df::Float64=4.0)

Build models from model equations with residual varainces R and degree of freedom for residual variance df defaulting to 4.0.

By default, all variabels in modelequations are fixed and factors. Set variables to be covariates or random using functions `setcovariate()orset_random()`.

#single-trait
model_equations = "BW = intercept + age + sex"
R               = 6.72
models          = build_model(model_equations,R);

#multi-trait
model_equations = "BW = intercept + age + sex
                   CW = intercept + litter";
R               = [6.72   24.84
                   24.84  708.41]
models          = build_model(model_equations,R);

source

JWAS.getEBV — Method.

getEBV(model::MME,genotypefile::AbstractString;header=true,separator=',')

Get estimated breeding values for individuals with genotypes defined by genotypefile. The genotype file format is same to that used in add_genotypes(). The same order for markers (columns) in genotypefile as in add_genotypes() is assumed for simplicity now. The file format is as follows,

Animal,marker1,marker2,marker3,marker4,marker5
A1,1,0,1,1,1
C3,1,2,0,1,0
B2,0,0,2,1,1

source

JWAS.getEBV — Method.

getEBV(model::MME,genotypes::Array{Float64,2})

Get estimated breeding values for individuals with genotypes. The genotypes are provided as an Array. The same order for markers (columns)in genotypes::Array{Float64,2} as in add_genotypes() is assumed for simplicity now.

source

JWAS.getEBV — Method.

getEBV(model::MME)

Get estimated breeding values for individuals defined by outputEBV() (defaulting to all genotyped individuals).

source

JWAS.outputEBV — Method.

outputEBV(model,IDs::Array;PEV=false)

Output estimated breeding values and prediction error variances (defaulting to false) for IDs.

source

JWAS.outputMCMCsamples — Method.

outputMCMCsamples(mme::MME,trmStr::AbstractString...)

Get MCMC samples for specific location parameters.

source

JWAS.runMCMC — Method.

runMCMC(model::MME,df::DataFrame;
        chain_length=1000,starting_value=false,burnin = 0,
        output_samples_frequency = 0,output_file="MCMC_samples",
        printout_model_info=true,printout_frequency=100,
        methods="conventional (no markers)",Pi=0.0,estimatePi=false,
        single_step_analysis= false,pedigree = false,
        missing_phenotypes=false,constraint=false,
        update_priors_frequency::Int64=0,
        outputEBV=true,output_PEV=false,output_heritability=false)

Run MCMC for Bayesian Linear Mixed Models with or without estimation of variance components.

Available methods include "conventional (no markers)", "RR-BLUP", "BayesB", "BayesC", "Bayesian Lasso", and "GBLUP".

Single step analysis is allowed if singlestepanalysis = true and pedigree is provided.

The starting_value can be provided as a vector of numbers for all location parameteres and marker effects, defaulting to 0.0s.

The first burnin iterations are discarded at the beginning of a MCMC chain of length chain_length.

Save MCMC samples every outputsamplesfrequency iterations, defaulting to false, to files outputsamplesfile, defaulting to MCMC_samples.txt. MCMC samples for hyperparametes (variance componets) and marker effects are saved by default if outputsamplesfrequency is provided. MCMC samples for location parametes can be saved using output_MCMC_samples(). Note that saving MCMC samples too frequently slows down the computation.

In Bayesian variable selection methods, Pi for single-trait analyses is a number; Pi for multi-trait analyses is a dictionary such as Pi=Dict([1.0; 1.0]=>0.7,[1.0; 0.0]=>0.1,[0.0; 1.0]=>0.1,[0.0; 0.0]=>0.1), defaulting to all markers have effects (0.0) in single-trait analysis and all markers have effects on all traits in multi-trait analysis. Pi is estimated if estimatePi = true

In multi-trait analysis, missing_phenotypes, defaulting to true, and constraint variance components, defaulting to false, are allowed. If constraint=true, constrain residual covariances between traits to be zeros.

Print out the model information in REPL if printout_model_info=true; print out the monte carlo mean in REPL with printout_frequency, defaulting to false.

Individual estimted breeding values (EBVs) are returned if outputEBV=true, defaulting to true. Heritability and genetic variances are returned if output_heritability=true, defaulting to false. Note that estimation of heritability is computaionally intensive.

source

JWAS.set_covariate — Method.

set_covariate(mme::MME,variables::AbstractString...)

set variables as covariates; mme is the output of function build_model().

#After running build_model, variabels age and year can be set to be covariates as
set_covariate(models,"age","year")
#or
set_covariate(models,"age year")

source

JWAS.set_random — Method.

set_random(mme::MME,randomStr::AbstractString,G;Vinv=0,names=[],df=4)

set variables as random effects, defaulting to i.i.d effects, with variances G with degree of freedom df, defaulting to 4.0.

the random effects are assumed to be i.i.d by default and it can be defined with any (inverse of) covariance structure Vinv with its index (row names) provided by names.

#single-trait (i.i.d randome effects)
model_equation  = "y = intercept + litter + sex"
model           = build_model(model_equation,R)
G               = 0.6
set_random(model,"litter",G)

#multi-trait (i.i.d randome effects)
model_equations = "BW = intercept + litter + sex
                   CW = intercept + litter + sex"
model           = build_model(model_equations,R);
G               = [3.72  1.84
                   1.84  3.41]
set_random(model,"litter",G)

#single-trait (randome effects with specific covariance structures)
model_equation  = "y = intercept + litter + sex"
model           = build_model(model_equation,R)
V               = [1.0  0.5 0.25
                   0.5  1.0 0.5
                   0.25 0.5 1.0]
G               = 0.6
set_random(model,"litter",G,Vinv=inv(V),names=[a1;a2;a3])

source

JWAS.set_random — Method.

set_random(mme::MME,randomStr::AbstractString,ped::Pedigree, G;df=4)

set variables as random polygenic effects with pedigree information ped and variances G with degree of freedom df, defaulting to 4.0.

#single-trait (example 1)
model_equation  = "y = intercept + age + animal"
model           = build_model(model_equation,R)
ped             = get_pedigree(pedfile)
G               = 1.6
set_random(model,"animal", ped, G)

#single-trait (example 2)
model_equation  = "y = intercept + age + animal + animal*age"
model           = build_model(model_equation,R)
ped             = get_pedigree(pedfile)
G               = [1.6   0.2
                   0.2  1.0]
set_random(model,"animal animal*age", ped,G)

#multi-trait
model_equations = "BW = intercept + age + sex + animal
                   CW = intercept + age + sex + animal"
model           = build_model(model_equations,R);
ped             = get_pedigree(pedfile);
G               = [6.72   2.84
                   2.84  8.41]
set_random(model,"animal",ped,G)

source

JWAS.showMME — Method.

showMME(mme::MME,df::DataFrame)

Show left-hand side and right-hand side of mixed model equations (no markers).

source

JWAS.solve — Method.

solve(mme::MME,df::DataFrame;solver="default",printout_frequency=100,tolerance = 0.000001,maxiter = 5000)

Solve the mixed model equations (no marker information) without estimating variance components.

Available solvers include default, Jacobi, GaussSeidel, and Gibbs sampler.

source

JWAS.add_term — Method.

add to model an extra term: imputation_residual

source

JWAS.getMME — Method.

Construct mixed model equations with

incidence matrix: X ; response : ySparse; left-hand side : mmeLhs ; right-hand side : mmeLhs ;

source

JWAS.get_outputX_others — Method.

get_outputX_others(model)

Make incidence matrices for effects involved in calculation of EBV including J, ϵ, pedTrmVec except marker covariates.

source

JWAS.PedModule.get_pedigree — Method.

get_pedigree(pedfile::AbstractString;header=false,separator=',')

Get pedigree informtion from a pedigree file with header (defaulting to false) and separator (defaulting to ,).

Pedigree file format:

a,0,0
c,a,b
d,a,c

source

–>